+++ wxc_sdk/telephony/dect_devices/__init__.py
@@ -11,7 +11,8 @@ from wxc_sdk.telephony.devices import AvailableMember
 
 __all__ = ['DECTNetworkModel', 'DECTNetworkDetail', 'BaseStationResult', 'BaseStationResponse', 'BaseStationsResponse',
            'DECTHandsetLine', 'Handset', 'BaseStationDetail', 'DECTHandsetItem', 'DECTHandsetList', 'UsageType',
-           'DectDevice', 'DECTDevicesApi']
+           'DectDevice', 'AddDECTHandset', 'AddDECTHandsetBulkError', 'AddDECTHandsetBulkResult',
+           'AddDECTHandsetBulkResponse', 'DECTDevicesApi']
 
 
 class DECTNetworkModel(str, Enum):
@@ -182,6 +183,37 @@ class DectDevice(ApiModel):
     number_of_line_ports: int
     #: Indicates number of supported registrations.
     number_of_registrations_supported: int
+
+
+class AddDECTHandset(ApiModel):
+    #: ID of the member on line1 of the handset. Members can be PEOPLE or PLACE.
+    line1_member_id: Optional[str] = None
+    #: ID of the member on line2 of the handset. Members can be PEOPLE, PLACE, or VIRTUAL_LINE.
+    line2_member_id: Optional[str] = None
+    #: Custom display name on the handset. Min and max length supported for the custom display name is 1 and 16
+    #: respectively.
+    custom_display_name: Optional[str] = None
+
+
+class AddDECTHandsetBulkError(ApiModel):
+    #: The error message.
+    message: Optional[str]
+    #: The error code.
+    error_code: Optional[str]
+
+
+class AddDECTHandsetBulkResult(ApiModel):
+    #: The status of the add handset request. 200 indicates the handset was added successfully.
+    status: Optional[int] = None
+    #: The error message if the add handset request failed.
+    error: Optional[AddDECTHandsetBulkError] = None
+
+
+class AddDECTHandsetBulkResponse(ApiModel):
+    #: The custom display name on the handset.
+    custom_display_name: Optional[str] = None
+    #: The result of the add handset request.
+    result: Optional[AddDECTHandsetBulkResult] = None
 
 
 class DECTDevicesApi(ApiChild, base='telephony/config'):
@@ -584,6 +616,12 @@ class DECTDevicesApi(ApiChild, base='telephony/config'):
         Adding a handset to a DECT network requires a full administrator auth token with a scope
         of `spark-admin:telephony_config_write`
 
+        Adding a DECT handset to a person with a Webex Calling Standard license will
+        disable Webex Calling across their Webex mobile, tablet, desktop, and browser applications.
+
+        Adding or removing handsets to the DECT network in less than 90 seconds may result
+        in base station not having the latest configuration until the base station is rebooted.
+
         :param location_id: Add handset in this location.
         :type location_id: str
         :param dect_network_id: A unique identifier for the DECT network.
@@ -612,6 +650,48 @@ class DECTDevicesApi(ApiChild, base='telephony/config'):
         body['customDisplayName'] = custom_display_name
         url = self.ep(f'locations/{location_id}/dectNetworks/{dect_network_id}/handsets')
         super().post(url, params=params, json=body)
+
+    def add_list_of_handsets(self, location_id: str, dect_network_id: str,
+                             items: list[AddDECTHandset],
+                             org_id: str = None) -> list[AddDECTHandsetBulkResponse]:
+        """
+        Add a List of Handsets to a DECT Network
+
+        Add a list of up to 50 handsets to a DECT network in a location.
+
+        A DECT network acts as a container that can support up to 1,000 lines across all handsets, with each handset
+        capable of handling up to two lines. Once the network is created, you can add bases, handsets, and assign
+        users or lines as needed.
+
+        Adding a list of handsets to a DECT network requires a full administrator auth token with a scope of
+        `spark-admin:telephony_config_write`.
+
+        Adding a DECT handset to a person with a Webex Calling Standard license will
+        disable Webex Calling across their Webex mobile, tablet, desktop, and browser applications.
+
+        Adding or removing handsets to the DECT network in less than 90 seconds may result
+        in base station not having the latest configuration until the base station is rebooted.
+
+        :param location_id: Add handsets in this location.
+        :type location_id: str
+        :param dect_network_id: A unique identifier for the DECT network.
+        :type dect_network_id: str
+        :param items: List of handsets that are to be added to the DECT network.
+        :type items: list[AddDECTHandset]
+        :param org_id: Add handsets in this organization.
+        :type org_id: str
+        :rtype: list[AddDECTHandsetBulkResponse]
+        """
+        params = {}
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = dict()
+        body['items'] = TypeAdapter(list[AddDECTHandset]).dump_python(items, mode='json', by_alias=True,
+                                                                      exclude_none=True)
+        url = self.ep(f'locations/{location_id}/dectNetworks/{dect_network_id}/handsets/bulk')
+        data = super().post(url, params=params, json=body)
+        r = TypeAdapter(list[AddDECTHandsetBulkResponse]).validate_python(data['items'])
+        return r
 
     def list_handsets(self, location_id: str, dect_network_id: str,
                       basestation_id: str = None, member_id: str = None,
@@ -767,6 +847,12 @@ class DECTDevicesApi(ApiChild, base='telephony/config'):
         can be of type PEOPLE, PLACE, or VIRTUAL_LINE.
 
         This API requires a full administrator auth token with a scope of `spark-admin:telephony_config_write`.
+
+        Deleting a DECT handset from a person with a Webex Calling Standard license will
+        enable Webex Calling across their Webex mobile, tablet, desktop, and browser applications.
+
+        Adding or removing handsets to the DECT network in less than 90 seconds may result
+        in base station not having the latest configuration until the base station is rebooted.
 
         :param location_id: Location containing the DECT network.
         :type location_id: str