+++ docs/user/changes.rst
@@ -7,6 +7,7 @@ Release history
 - feat: new API :attr:`api.telephony.hotdesk <wxc_sdk.telephony.TelephonyApi.hotdesk>`
 - feat: new parameter carrier_id for :meth:`api.telephony.location.number.add <wxc_sdk.telephony.location.numbers.LocationNumbersApi.add>`
 - feat: new API :attr:`api.person_settings.feature_access <wxc_sdk.person_settings.PersonSettingsApi.feature_access>`
+- feat: new endpoint: :meth:`api.telephony.dect_devices.add_list_of_handsets <wxc_sdk.telephony.dect_devices.DECTDevicesApi.add_list_of_handsets>`
 
 1.25.0
 ------